Refactor MCMS Go tests
