+++ relayer/txm/txm_multisig_deploy_ccip_local_test.go
@@ -118,7 +118,7 @@ func chunkMetadata(metadata []byte, chunkSize int) []TimelockOperation {
 		chunk := metadata[i:end]
 
 		// Create operation for this chunk
-		metadataData, _ := serializeStageCodeChunkParams(chunk, []uint16{}, [][]byte{}, []byte{})
+		metadataData, _ := SerializeStageCodeChunkParams(chunk, []uint16{}, [][]byte{}, []byte{})
 		metadataOp := TimelockOperation{
 			Target:       deployMcmsAccount,
 			ModuleName:   "mcms_deployer",
@@ -144,7 +144,7 @@ func chunkBytecode(bytecodeModules [][]byte, chunkSize int) []TimelockOperation 
 			chunk := moduleBytes[i:end]
 
 			// Create operation for this chunk
-			moduleData, _ := serializeStageCodeChunkParams([]byte{}, []uint16{uint16(moduleIdx)}, [][]byte{chunk}, []byte{})
+			moduleData, _ := SerializeStageCodeChunkParams([]byte{}, []uint16{uint16(moduleIdx)}, [][]byte{chunk}, []byte{})
 			moduleOp := TimelockOperation{
 				Target:       deployMcmsAccount,
 				ModuleName:   "mcms_deployer",
@@ -163,7 +163,7 @@ func scheduleAndExecuteOperations(
 	logger logger.Logger,
 	txm *AptosTxm,
 	operations []TimelockOperation,
-	bypasserSigners []Signer,
+	proposerSigners []Signer,
 	deployerAddress string,
 	deployerPublicKeyHex string,
 	role uint8,
@@ -187,20 +187,20 @@ func scheduleAndExecuteOperations(
 	ops := deployTimelockOpToMultipleOps(operations, predecessor, salt, delay, role, chainID)
 
 	// Generate merkle tree and set root
-	merkleTree, err := generateMerkleTree(ops, rootMetadata)
+	merkleTree, err := GenerateMerkleTree(ops, rootMetadata)
 	if err != nil {
 		t.Fatalf("Failed to generate merkle tree: %v", err)
 	}
 
-	setRootId := enqueueSetRoot(t, logger, merkleTree, txm, rootMetadata, bypasserSigners,
+	setRootId := EnqueueSetRoot(t, logger, merkleTree, txm, rootMetadata, proposerSigners,
 		deployerAddress, deployerPublicKeyHex, deployMcmsAddress)
 	waitForTxmId(t, txm, setRootId, time.Second*30)
 
 	// Schedule each operation through MCMS
 	for i, op := range operations {
-		proof := merkleTree.getProof(i + 1)
-		require.True(t, merkleTree.verifyProof(proof, hashOp(&ops[i])))
-		txId := scheduleSingleOperationAsDeployer(t, logger, txm, deployMcmsAccount,
+		proof := merkleTree.GetProof(i + 1)
+		require.True(t, merkleTree.VerifyProof(proof, HashOp(&ops[i])))
+		txId := ScheduleSingleOperationAsDeployer(t, logger, txm, deployMcmsAccount,
 			deployerAddress, deployerPublicKeyHex, []TimelockOperation{op},
 			predecessor, salt, delay, role, chainID, ops[i].Nonce, proof)
 		waitForTxmId(t, txm, txId, time.Second*30)
@@ -211,7 +211,7 @@ func scheduleAndExecuteOperations(
 
 	// Execute each operation
 	for i := 0; i < len(operations); i++ {
-		txId := executeBatchOperations(t, txm, deployMcmsAddress, deployerAddress,
+		txId := ExecuteBatchOperations(t, txm, deployMcmsAddress, deployerAddress,
 			deployerPublicKeyHex, []TimelockOperation{operations[i]}, predecessor, salt)
 		waitForTxmId(t, txm, txId, time.Second*30)
 	}
@@ -220,10 +220,10 @@ func scheduleAndExecuteOperations(
 // runDeployMCMSAndCCIPInChunks contains the core logic for deploying
 // the CCIP contract via MCMS chunks.
 func runDeployMCMSAndCCIPInChunks(t *testing.T, logger logger.Logger, rpcURL string, keystore loop.Keystore, accounts []Account) {
-	// Generate bypasser signers for the test
-	bypasserSigners := generateSigners(t, 3)
-	for _, signer := range bypasserSigners {
-		logger.Debugw("Generated bypasser signer", "address", hex.EncodeToString(signer.address))
+	// Generate proposer signers for the test
+	proposerSigners := GenerateSigners(t, 3)
+	for _, signer := range proposerSigners {
+		logger.Debugw("Generated proposer signer", "address", hex.EncodeToString(signer.Address))
 	}
 
 	deployer := accounts[0]
@@ -278,9 +278,9 @@ func runDeployMCMSAndCCIPInChunks(t *testing.T, logger logger.Logger, rpcURL str
 	require.NoError(t, err)
 	waitForTxmId(t, txm, deployId, time.Second*30)
 
-	role := uint8(BYPASSER_ROLE)
-	setupInitialConfigAsDeployer(t, logger, txm, role, bypasserSigners, deployerAddress, deployerPublicKeyHex, true, deployMcmsAddress)
-	transferOwnership(t, txm, deployerAddress, deployerPublicKeyHex, deployMcmsAddress)
+	role := uint8(PROPOSER_ROLE)
+	SetupInitialConfigAsDeployer(t, logger, txm, role, proposerSigners, deployerAddress, deployerPublicKeyHex, true, deployMcmsAddress)
+	TransferOwnership(t, txm, deployerAddress, deployerPublicKeyHex, deployMcmsAddress)
 
 	// Create a unique seed for the new object owner
 	newOwnerSeed := make([]byte, 32)
@@ -292,7 +292,7 @@ func runDeployMCMSAndCCIPInChunks(t *testing.T, logger logger.Logger, rpcURL str
 	require.NoError(t, err, "Failed to serialize newOwnerSeed for view call")
 
 	// Call mcms::mcms_registry::get_new_code_object_address
-	codeObjectAddressPayload := createViewPayload(deployMcmsAddress, "mcms_registry", "get_new_code_object_address",
+	codeObjectAddressPayload := CreateViewPayload(deployMcmsAddress, "mcms_registry", "get_new_code_object_address",
 		[]aptos.TypeTag{}, [][]byte{seedArg})
 	viewResult, err := deployClient.View(codeObjectAddressPayload)
 	require.NoError(t, err, "Failed to call get_new_code_object_address view function")
@@ -322,7 +322,7 @@ func runDeployMCMSAndCCIPInChunks(t *testing.T, logger logger.Logger, rpcURL str
 	moduleOps := chunkBytecode(compileResult.BytecodeModules, CHUNK_SIZE)
 
 	// 4. Create final publish operation
-	publishData, err := serializeStageCodeChunkParams([]byte{}, []uint16{}, [][]byte{}, newOwnerSeed)
+	publishData, err := SerializeStageCodeChunkParams([]byte{}, []uint16{}, [][]byte{}, newOwnerSeed)
 	require.NoError(t, err)
 	publishOp := TimelockOperation{
 		Target:       deployMcmsAccount,
@@ -340,10 +340,10 @@ func runDeployMCMSAndCCIPInChunks(t *testing.T, logger logger.Logger, rpcURL str
 	// Schedule and execute operations
 	delay := uint64(TEST_DELAY)
 	salt := []byte{} // Empty salt for initial deployment
-	scheduleAndExecuteOperations(t, logger, txm, allOperations, bypasserSigners, deployerAddress, deployerPublicKeyHex, role, deployChainIdBig, delay, salt)
+	scheduleAndExecuteOperations(t, logger, txm, allOperations, proposerSigners, deployerAddress, deployerPublicKeyHex, role, deployChainIdBig, delay, salt)
 
 	// Verify deployment was successful
-	onrampTypeAndVersionPayload := createViewPayload(ccipObjectAddress.String(), "onramp", "type_and_version",
+	onrampTypeAndVersionPayload := CreateViewPayload(ccipObjectAddress.String(), "onramp", "type_and_version",
 		[]aptos.TypeTag{}, [][]byte{})
 	result, err := deployClient.View(onrampTypeAndVersionPayload)
 	require.NoError(t, err)
@@ -425,11 +425,11 @@ func runDeployMCMSAndCCIPInChunks(t *testing.T, logger logger.Logger, rpcURL str
 
 	// Schedule and execute upgrade operations (with unique salt)
 	upgradeSalt := []byte("upgrade") // Use a different salt for the upgrade
-	scheduleAndExecuteOperations(t, logger, txm, allUpgradeOps, bypasserSigners,
+	scheduleAndExecuteOperations(t, logger, txm, allUpgradeOps, proposerSigners,
 		deployerAddress, deployerPublicKeyHex, role, deployChainIdBig, delay, upgradeSalt)
 
 	// Verify upgrade was successful
-	updatedVersionPayload := createViewPayload(ccipObjectAddress.String(), "onramp", "type_and_version",
+	updatedVersionPayload := CreateViewPayload(ccipObjectAddress.String(), "onramp", "type_and_version",
 		[]aptos.TypeTag{}, [][]byte{})
 	updatedResult, err := deployClient.View(updatedVersionPayload)
 	require.NoError(t, err)
@@ -468,7 +468,7 @@ func deployTimelockOpToMultipleOps(tlOps []TimelockOperation, predecessor []byte
 
 	for i, tlOp := range tlOps {
 		// For each operation, serialize its parameters individually
-		serializedData, err := serializeScheduleBatchParams([]TimelockOperation{tlOp}, predecessor, salt, delay)
+		serializedData, err := SerializeScheduleBatchParams([]TimelockOperation{tlOp}, predecessor, salt, delay)
 		if err != nil {
 			panic(fmt.Sprintf("failed to serialize operation: %v", err))
 		}
