+++ relayer/txm/txm_multisig_local_test.go
@@ -3,14 +3,10 @@
 package txm
 
 import (
-	"bytes"
 	"context"
-	"crypto/ecdsa"
 	"crypto/ed25519"
 	"crypto/rand"
-	"encoding/binary"
 	"encoding/hex"
-	"errors"
 	"fmt"
 	"math/big"
 	"path/filepath"
@@ -22,7 +18,6 @@ import (
 	aptosapi "github.com/aptos-labs/aptos-go-sdk/api"
 	"github.com/aptos-labs/aptos-go-sdk/bcs"
 	aptoscrypto "github.com/aptos-labs/aptos-go-sdk/crypto"
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/go-viper/mapstructure/v2"
 	"github.com/google/uuid"
@@ -42,18 +37,6 @@ type Account struct {
 	accountAddress aptos.AccountAddress
 }
 
-type Signer struct {
-	// ethereum address
-	address    []byte
-	privateKey *ecdsa.PrivateKey
-}
-
-const TEST_DELAY = uint64(1)
-
-const BYPASSER_ROLE = 0
-const CANCELLOR_ROLE = 1
-const PROPOSER_ROLE = 2
-
 var (
 	client      *aptos.NodeClient
 	mcmsAccount aptos.AccountAddress
@@ -64,19 +47,16 @@ var (
 	preOpCount  uint64 = 0
 )
 
-// Add package-level getNextNonce function
 func getNextNonce() uint64 {
 	nonce := nextNonce
 	nextNonce++
 	return nonce
 }
 
-// Add package-level getPreOpCount function
 func getPreOpCount() uint64 {
 	return postOpCount
 }
 
-// Add package-level getCumulativePostOpCount function
 func getCumulativePostOpCount(numOps uint64) uint64 {
 	postOpCount += numOps
 	return postOpCount
@@ -131,14 +111,14 @@ func TestMultisigLocal(t *testing.T) {
 
 func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore loop.Keystore, accounts []Account, skipEOATest bool) {
 	// Generate bypasser and proposer signers
-	bypasserSigners := generateSigners(t, 3)
-	proposerSigners := generateSigners(t, 3)
+	bypasserSigners := GenerateSigners(t, 3)
+	proposerSigners := GenerateSigners(t, 3)
 	// Log generated signers
 	for _, signer := range bypasserSigners {
-		logger.Debugw("Generated bypasser signer", "address", hex.EncodeToString(signer.address))
+		logger.Debugw("Generated bypasser signer", "address", hex.EncodeToString(signer.Address))
 	}
 	for _, signer := range proposerSigners {
-		logger.Debugw("Generated proposer signer", "address", hex.EncodeToString(signer.address))
+		logger.Debugw("Generated proposer signer", "address", hex.EncodeToString(signer.Address))
 	}
 
 	deployer := accounts[0]
@@ -179,7 +159,7 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 	deployId := uuid.New().String()
 	err = txm.Enqueue(
 		deployId,
-		getSampleTxMetadata(),
+		GetSampleTxMetadata(),
 		deployerAddress,
 		deployerPublicKeyHex,
 		"0x1::resource_account::create_resource_account_and_publish_package",
@@ -189,7 +169,7 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 		/* simulateTx= */ true,
 	)
 	require.NoError(t, err)
-	waitForTxmId(t, txm, deployId, time.Second*30)
+	WaitForTxmId(t, txm, deployId, time.Second*30)
 
 	logger.Infow("published module", "deployerAddress", deployerAddress, "mcmsAddress", mcmsAddress)
 
@@ -223,9 +203,9 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 
 	// Call set_config to set signers
 	role := uint8(PROPOSER_ROLE)
-	setupInitialConfigAsDeployer(t, logger, txm, uint8(BYPASSER_ROLE), bypasserSigners, deployerAddress, deployerPublicKeyHex, false, mcmsAddress)
+	SetupInitialConfigAsDeployer(t, logger, txm, uint8(BYPASSER_ROLE), bypasserSigners, deployerAddress, deployerPublicKeyHex, false, mcmsAddress)
 	// Setup role for Proposer too
-	setupInitialConfigAsDeployer(t, logger, txm, role, proposerSigners, deployerAddress, deployerPublicKeyHex, false, mcmsAddress)
+	SetupInitialConfigAsDeployer(t, logger, txm, role, proposerSigners, deployerAddress, deployerPublicKeyHex, false, mcmsAddress)
 
 	functionOneArg1 := "hello"
 	functionOneArg2 := []byte{5, 4, 3, 2, 1}
@@ -233,14 +213,14 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 	functionTwoArg2 := big.NewInt(42)
 
 	// function_one(arg1: String, arg2: vector<u8>)
-	functionOneParamBytes := getFunctionOneParamBytes(t, functionOneArg1, functionOneArg2)
+	functionOneParamBytes := GetFunctionOneParamBytes(t, functionOneArg1, functionOneArg2)
 
 	// function_two(arg1: address, arg2: u128)
-	functionTwoParamBytes := getFunctionTwoParamBytes(t, functionTwoArg1, functionTwoArg2)
+	functionTwoParamBytes := GetFunctionTwoParamBytes(t, functionTwoArg1, functionTwoArg2)
 
 	t.Run("TransferOwnershipAndExecuteMCMSDeployedUserContract", func(t *testing.T) {
 		role := uint8(PROPOSER_ROLE)
-		transferOwnership(t, txm, deployerAddress, deployerPublicKeyHex, mcmsAddress)
+		TransferOwnership(t, txm, deployerAddress, deployerPublicKeyHex, mcmsAddress)
 
 		newOwnerSeed := []byte("mcms_user_contract")
 		codeObjectOwnerAccount := mcmsAccount.ResourceAccount(append([]byte("CHAINLINK_MCMS_NEW_OBJECT_REGISTRATION"), newOwnerSeed...))
@@ -317,26 +297,26 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 			OverridePreviousRoot: false,
 		}
 
-		ops := timelockOpToMulitpleOps(scheduleTimelockOps, predecessor, salt, delay, role, chainIdBig)
+		ops := TimelockOpToMultipleOps(scheduleTimelockOps, predecessor, salt, delay, role, chainIdBig, mcmsAccount, getNextNonce)
 
 		// Log hashRootMetadata and contents
-		metadataHash := hashRootMetadata(rootMetadata)
+		metadataHash := HashRootMetadata(rootMetadata)
 		logger.Debugw("hashRootMetadata", "value", hex.EncodeToString(metadataHash[:]))
 		logger.Debugw("ops", "value", ops)
 
-		merkleTree, err := generateMerkleTree(ops, rootMetadata)
+		merkleTree, err := GenerateMerkleTree(ops, rootMetadata)
 		require.NoError(t, err)
 
-		setRootId := enqueueSetRoot(t, logger, merkleTree, txm, rootMetadata, proposerSigners, deployerAddress, deployerPublicKeyHex, mcmsAddress)
-		waitForTxmId(t, txm, setRootId, time.Second*30)
+		setRootId := EnqueueSetRoot(t, logger, merkleTree, txm, rootMetadata, proposerSigners, deployerAddress, deployerPublicKeyHex, mcmsAddress)
+		WaitForTxmId(t, txm, setRootId, time.Second*30)
 
 		// Schedule operations through MCMS -> MCMS Timelock
 		for i, op := range scheduleTimelockOps {
-			proof := merkleTree.getProof(i + 1)
-			require.True(t, merkleTree.verifyProof(proof, hashOp(&ops[i])))
-			txId := scheduleSingleOperationAsDeployer(t, logger, txm, mcmsAccount, deployerAddress, deployerPublicKeyHex,
+			proof := merkleTree.GetProof(i + 1)
+			require.True(t, merkleTree.VerifyProof(proof, HashOp(&ops[i])))
+			txId := ScheduleSingleOperationAsDeployer(t, logger, txm, mcmsAccount, deployerAddress, deployerPublicKeyHex,
 				[]TimelockOperation{op}, predecessor, salt, delay, role, chainIdBig, ops[i].Nonce, proof)
-			waitForTxmId(t, txm, txId, time.Second*30)
+			WaitForTxmId(t, txm, txId, time.Second*30)
 		}
 
 		// Wait for timelock delay
@@ -372,9 +352,9 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 
 		// Here we execute each op one by one, to prevent EXECUTION_LIMIT_REACHED
 		for i := 0; i < len(executeTimelockOps); i++ {
-			txId := executeBatchOperations(t, txm, mcmsAddress, deployerAddress, deployerPublicKeyHex,
+			txId := ExecuteBatchOperations(t, txm, mcmsAddress, deployerAddress, deployerPublicKeyHex,
 				[]TimelockOperation{executeTimelockOps[i]}, predecessor, salt)
-			waitForTxmId(t, txm, txId, time.Second*30)
+			WaitForTxmId(t, txm, txId, time.Second*30)
 		}
 
 		// Verify final state
@@ -432,7 +412,7 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 			// deploy and initialize mcms user module
 			err = txm.Enqueue(
 				uuid.New().String(),
-				getSampleTxMetadata(),
+				GetSampleTxMetadata(),
 				mcmsUserDeployerAddress,
 				mcmsUserDeployerPublicKeyHex,
 				"0x1::code::publish_package_txn",
@@ -488,27 +468,27 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 				OverridePreviousRoot: false,
 			}
 
-			op := timelockOpsToOp(scheduleTimelockOps, predecessor, salt, delay, role, chainIdBig, mcmsAccount, getNextNonce)
+			op := TimelockOpsToOp(scheduleTimelockOps, predecessor, salt, delay, role, chainIdBig, mcmsAccount, getNextNonce)
 
-			merkleTree, err := generateMerkleTree([]Op{op}, rootMetadata)
+			merkleTree, err := GenerateMerkleTree([]Op{op}, rootMetadata)
 			require.NoError(t, err)
 
-			setRootId := enqueueSetRoot(t, logger, merkleTree, txm, rootMetadata, proposerSigners, deployerAddress, deployerPublicKeyHex, mcmsAddress)
-			waitForTxmId(t, txm, setRootId, time.Second*30)
+			setRootId := EnqueueSetRoot(t, logger, merkleTree, txm, rootMetadata, proposerSigners, deployerAddress, deployerPublicKeyHex, mcmsAddress)
+			WaitForTxmId(t, txm, setRootId, time.Second*30)
 
-			proof := merkleTree.getProof(1)
+			proof := merkleTree.GetProof(1)
 
 			// Calculate the hash of our single operation (for verification)
-			opHash := hashOp(&op)
+			opHash := HashOp(&op)
 			logger.Debugw("Leaf hash for operation", "hash", hex.EncodeToString(opHash[:]))
 
 			// Verify that the proof is valid for our operation
-			require.True(t, merkleTree.verifyProof(proof, opHash), "Failed to verify merkle proof for the batched operation")
+			require.True(t, merkleTree.VerifyProof(proof, opHash), "Failed to verify merkle proof for the batched operation")
 
 			// Schedule all operations in a single execute call, passing the complete proof
-			txId := scheduleBatchOperationsAsDeployer(t, logger, txm, mcmsAccount, deployerAddress, deployerPublicKeyHex,
+			txId := ScheduleBatchOperationsAsDeployer(t, logger, txm, mcmsAccount, deployerAddress, deployerPublicKeyHex,
 				scheduleTimelockOps, predecessor, salt, delay, role, chainIdBig, op.Nonce, proof)
-			waitForTxmId(t, txm, txId, time.Second*30)
+			WaitForTxmId(t, txm, txId, time.Second*30)
 
 			for {
 				queueLen, unconfirmedLen := txm.InflightCount()
@@ -535,8 +515,8 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 				},
 			}
 
-			txId = executeBatchOperations(t, txm, mcmsAddress, deployerAddress, deployerPublicKeyHex, executeTimelockOps, predecessor, salt)
-			waitForTxmId(t, txm, txId, time.Second*30)
+			txId = ExecuteBatchOperations(t, txm, mcmsAddress, deployerAddress, deployerPublicKeyHex, executeTimelockOps, predecessor, salt)
+			WaitForTxmId(t, txm, txId, time.Second*30)
 
 			// check that user contract state was updated
 			{
@@ -594,7 +574,7 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 
 		// Addresses are already sorted
 		for _, signer := range proposerSigners {
-			signerAddresses = append(signerAddresses, signer.address)
+			signerAddresses = append(signerAddresses, signer.Address)
 			signerGroups = append(signerGroups, 0)
 		}
 		groupQuorums[0] = 2
@@ -604,11 +584,11 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 		})
 		require.NoError(t, err)
 
-		proposerMultisig := getMultisigForRole(t, mcmsAddress, client, role)
+		proposerMultisig := GetMultisigForRole(t, mcmsAddress, client, role)
 		logger.Debugw("Proposer multisig", "multisig", proposerMultisig.String())
 
 		paramValues := [][]byte{[]byte{role}}
-		viewPayload := createViewPayload(mcmsAddress, "mcms", "get_config", []aptos.TypeTag{}, paramValues)
+		viewPayload := CreateViewPayload(mcmsAddress, "mcms", "get_config", []aptos.TypeTag{}, paramValues)
 		data, err := client.View(viewPayload)
 		require.NoError(t, err)
 		logger.Debugw("Initial config data", "data", data)
@@ -620,7 +600,7 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 
 		// Now to call mcms::set_config, we must schedule through `mcms::schedule_batch`
 		// We change config to remove a signer
-		setConfigData, err := serializeSetConfig(role, SignerConfig{
+		setConfigData, err := SerializeSetConfig(role, SignerConfig{
 			// Remove the last signer
 			Addresses:    signerAddresses[:len(signerAddresses)-1],
 			Groups:       signerGroups[:len(signerGroups)-1],
@@ -656,13 +636,13 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 			PostOpCount:          getPreOpCount() + 1, // We execute this all in one OP
 			OverridePreviousRoot: false,
 		}
-		op := timelockOpsToOp(scheduleTimelockOps, predecessor, salt, delay, role, chainIdBig, mcmsAccount, getNextNonce)
+		op := TimelockOpsToOp(scheduleTimelockOps, predecessor, salt, delay, role, chainIdBig, mcmsAccount, getNextNonce)
 
-		merkleTree, err := generateMerkleTree([]Op{op}, rootMetadata)
+		merkleTree, err := GenerateMerkleTree([]Op{op}, rootMetadata)
 		require.NoError(t, err)
 
-		setRootId := enqueueSetRoot(t, logger, merkleTree, txm, rootMetadata, proposerSigners, deployerAddress, deployerPublicKeyHex, mcmsAddress)
-		waitForTxmId(t, txm, setRootId, time.Second*30)
+		setRootId := EnqueueSetRoot(t, logger, merkleTree, txm, rootMetadata, proposerSigners, deployerAddress, deployerPublicKeyHex, mcmsAddress)
+		WaitForTxmId(t, txm, setRootId, time.Second*30)
 
 		// The proper approach is to schedule all operations in a single execute call.
 		// Since timelockOpsToOps now creates a single Op that contains all operations,
@@ -672,19 +652,19 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 
 		// Get the complete proof for leaf index 1 (our single operation containing all operations)
 		// This proof contains all sibling nodes needed to verify the path from leaf to root
-		proof := merkleTree.getProof(1)
+		proof := merkleTree.GetProof(1)
 
 		// Calculate the hash of our single operation (for verification)
-		opHash := hashOp(&op)
+		opHash := HashOp(&op)
 		logger.Debugw("Leaf hash for operation", "hash", hex.EncodeToString(opHash[:]))
 
 		// Verify that the proof is valid for our operation
-		require.True(t, merkleTree.verifyProof(proof, opHash), "Failed to verify merkle proof for the batched operation")
+		require.True(t, merkleTree.VerifyProof(proof, opHash), "Failed to verify merkle proof for the batched operation")
 
 		// Schedule all operations in a single execute call, passing the complete proof
-		txId := scheduleBatchOperationsAsDeployer(t, logger, txm, mcmsAccount, deployerAddress, deployerPublicKeyHex,
+		txId := ScheduleBatchOperationsAsDeployer(t, logger, txm, mcmsAccount, deployerAddress, deployerPublicKeyHex,
 			scheduleTimelockOps, predecessor, salt, delay, role, chainIdBig, op.Nonce, proof)
-		waitForTxmId(t, txm, txId, time.Second*30)
+		WaitForTxmId(t, txm, txId, time.Second*30)
 
 		// Wait for test delay
 		time.Sleep(time.Duration(TEST_DELAY) * time.Second)
@@ -706,8 +686,8 @@ func runMultisigTest(t *testing.T, logger logger.Logger, rpcURL string, keystore
 		}
 
 		// Execute all operations in a single batch call to execute_batch
-		txId = executeBatchOperations(t, txm, mcmsAddress, deployerAddress, deployerPublicKeyHex, executeTimelockOps, predecessor, salt)
-		waitForTxmId(t, txm, txId, time.Second*30)
+		txId = ExecuteBatchOperations(t, txm, mcmsAddress, deployerAddress, deployerPublicKeyHex, executeTimelockOps, predecessor, salt)
+		WaitForTxmId(t, txm, txId, time.Second*30)
 
 		// Get number of signers after set_config to 2 signers (3 initially)
 		postSetConfigData, postSetConfigErr := client.View(viewPayload)
@@ -748,279 +728,6 @@ func compileMcmsUserContract(t *testing.T, packageAddress, mcmsAddress aptos.Acc
 	return compileResult.PackageMetadata, compileResult.BytecodeModules[0]
 }
 
-type RootMetadata struct {
-	Role                 uint8
-	ChainID              *big.Int
-	MultiSig             aptos.AccountAddress
-	PreOpCount           uint64
-	PostOpCount          uint64
-	OverridePreviousRoot bool
-}
-
-type Op struct {
-	Role         uint8
-	ChainID      *big.Int
-	MultiSig     aptos.AccountAddress
-	Nonce        uint64
-	To           aptos.AccountAddress
-	ModuleName   string
-	FunctionName string
-	Data         []byte
-}
-
-func enqueueSetRoot(
-	t *testing.T, logger logger.Logger, merkleTree MerkleTree, txm *AptosTxm, rootMetadata RootMetadata,
-	signers []Signer, deployerAddress string, deployerPublicKeyHex string, mcmsAddress string,
-) string {
-	rootHash := merkleTree.getRoot()
-
-	// Set validUntil to be the current UTC timestamp + 1 day
-	validUntil := uint64(time.Now().UTC().Add(1 * 24 * time.Hour).Unix())
-
-	signedHash := calculateSignedHash(rootHash, validUntil)
-	signatures := generateSignatures(t, signers, signedHash)
-
-	// The first leaf is the metadata
-	metadataProof := merkleTree.getProof(0)
-
-	require.True(t, merkleTree.verifyProof(metadataProof, hashRootMetadata(rootMetadata)))
-	// log hashrootmetadata
-	logger.Debugw("rootMetadata", "rootMetadata", rootMetadata)
-	logger.Debugw("hashRootMetadata(rootMetadata)", "hashRootMetadata(rootMetadata)", hashRootMetadata(rootMetadata))
-
-	// Log all values needed for Move e2e test
-	logger.Debugw("============= BEGIN VALUES FOR MOVE E2E TEST =============")
-	logger.Debugw("Root Hash", "value", hex.EncodeToString(rootHash[:]))
-	logger.Debugw("Valid Until", "value", validUntil)
-	logger.Debugw("Signed Hash", "value", hex.EncodeToString(signedHash[:]))
-
-	signaturesHex := make([]string, len(signatures))
-	for i, sig := range signatures {
-		signaturesHex[i] = hex.EncodeToString(sig)
-	}
-	logger.Debugw("Signatures", "values", signaturesHex)
-
-	signerAddressesHex := make([]string, len(signers))
-	for i, signer := range signers {
-		signerAddressesHex[i] = hex.EncodeToString(signer.address)
-	}
-	logger.Debugw("Signer Addresses", "values", signerAddressesHex)
-
-	metadataProofHex := make([]string, len(metadataProof))
-	for i, p := range metadataProof {
-		metadataProofHex[i] = hex.EncodeToString(p[:])
-	}
-	logger.Debugw("Metadata Proof", "values", metadataProofHex)
-
-	// Log all leaves
-	leavesHex := make([]string, len(merkleTree))
-	for i := 0; i < len(merkleTree); i++ {
-		leavesHex[i] = hex.EncodeToString(merkleTree[i][:])
-	}
-	logger.Debugw("All Merkle Tree Nodes (including internal nodes)", "values", leavesHex)
-
-	// Log just the leaf nodes (metadata + ops)
-	leafNodesHex := make([]string, 0)
-	for i := 0; i < len(merkleTree) && i < int(rootMetadata.PostOpCount-rootMetadata.PreOpCount+1); i++ {
-		leafNodesHex = append(leafNodesHex, hex.EncodeToString(merkleTree[i][:]))
-	}
-	logger.Debugw("Leaf Nodes Only (metadata + ops)", "values", leafNodesHex)
-
-	// Log root metadata and its hash
-	metadataHash := hashRootMetadata(rootMetadata)
-	logger.Debugw("Root Metadata Hash", "value", hex.EncodeToString(metadataHash[:]))
-	logger.Debugw("Root Metadata Values",
-		"Role", rootMetadata.Role,
-		"ChainID", rootMetadata.ChainID,
-		"MultiSig", rootMetadata.MultiSig.String(),
-		"PreOpCount", rootMetadata.PreOpCount,
-		"PostOpCount", rootMetadata.PostOpCount,
-		"OverridePreviousRoot", rootMetadata.OverridePreviousRoot)
-
-	logger.Debugw("============= END VALUES FOR MOVE E2E TEST =============")
-
-	setRootId := uuid.New().String()
-
-	err := txm.Enqueue(
-		setRootId,
-		getSampleTxMetadata(),
-		deployerAddress,
-		deployerPublicKeyHex,
-		mcmsAddress+"::mcms::set_root",
-		[]string{},
-		[]string{
-			"u8",
-			"vector<u8>",
-			"u64",
-			"u256",
-			"address",
-			"u64",
-			"u64",
-			"bool",
-			"vector<vector<u8>>",
-			"vector<vector<u8>>",
-		},
-		[]any{
-			rootMetadata.Role,
-			rootHash,
-			validUntil,
-			rootMetadata.ChainID,
-			rootMetadata.MultiSig,
-			rootMetadata.PreOpCount,
-			rootMetadata.PostOpCount,
-			rootMetadata.OverridePreviousRoot,
-			metadataProof,
-			signatures,
-		},
-		/* simulateTx= */ true,
-	)
-	require.NoError(t, err)
-
-	return setRootId
-}
-
-func generateMerkleTree(ops []Op, rootMetadata RootMetadata) (MerkleTree, error) {
-	leaves := make([][32]byte, len(ops)+1)
-	leaves[0] = hashRootMetadata(rootMetadata)
-	for i, op := range ops {
-		leaves[i+1] = hashOp(&op)
-	}
-	return newMerkleTree(leaves)
-}
-
-func hashRootMetadata(metadata RootMetadata) common.Hash {
-	MANY_CHAIN_MULTI_SIG_DOMAIN_SEPARATOR_METADATA := crypto.Keccak256([]byte("MANY_CHAIN_MULTI_SIG_DOMAIN_SEPARATOR_METADATA_APTOS"))
-	ser := bcs.Serializer{}
-	ser.FixedBytes(MANY_CHAIN_MULTI_SIG_DOMAIN_SEPARATOR_METADATA)
-	ser.U8(metadata.Role)
-	ser.U256(*metadata.ChainID)
-	ser.Struct(&metadata.MultiSig)
-	ser.U64(metadata.PreOpCount)
-	ser.U64(metadata.PostOpCount)
-	ser.Bool(metadata.OverridePreviousRoot)
-
-	return crypto.Keccak256Hash(ser.ToBytes())
-}
-
-func hashOp(op *Op) common.Hash {
-	MANY_CHAIN_MULTI_SIG_DOMAIN_SEPARATOR_OP := crypto.Keccak256([]byte("MANY_CHAIN_MULTI_SIG_DOMAIN_SEPARATOR_OP_APTOS"))
-	ser := bcs.Serializer{}
-	ser.FixedBytes(MANY_CHAIN_MULTI_SIG_DOMAIN_SEPARATOR_OP)
-	ser.U8(op.Role)
-	ser.U256(*op.ChainID)
-	ser.Struct(&op.MultiSig)
-	ser.U64(op.Nonce)
-	ser.Struct(&op.To)
-	ser.WriteString(op.ModuleName)
-	ser.WriteString(op.FunctionName)
-	ser.WriteBytes(op.Data)
-
-	return crypto.Keccak256Hash(ser.ToBytes())
-}
-
-func hashPair(left, right [32]byte) [32]byte {
-	if bytes.Compare(left[:], right[:]) < 0 {
-		return crypto.Keccak256Hash(left[:], right[:])
-	}
-	return crypto.Keccak256Hash(right[:], left[:])
-}
-
-type MerkleTree [][32]byte
-
-func (mt MerkleTree) getRoot() [32]byte {
-	return mt[len(mt)-1]
-}
-
-func newMerkleTree(leaves [][32]byte) (MerkleTree, error) {
-	if len(leaves) == 0 {
-		return nil, errors.New("empty leaf set")
-	}
-
-	// Calculate the next power of 2
-	leafCount := len(leaves)
-	treeSize := 1
-	for treeSize < leafCount {
-		treeSize *= 2
-	}
-
-	// Create a new slice with the correct size
-	paddedLeaves := make([][32]byte, treeSize)
-	copy(paddedLeaves, leaves)
-
-	// Fill the rest with zero leaves
-	zeroLeaf := [32]byte{}
-	for i := leafCount; i < treeSize; i++ {
-		paddedLeaves[i] = zeroLeaf
-	}
-
-	tree := make(MerkleTree, treeSize)
-	copy(tree, paddedLeaves)
-
-	index := 0
-	for levelSize := treeSize; levelSize > 1; levelSize /= 2 {
-		for i := index; i < index+levelSize; i += 2 {
-			tree = append(tree, hashPair(tree[i], tree[i+1]))
-		}
-		index += levelSize
-	}
-
-	return tree, nil
-}
-
-func (mt MerkleTree) getProof(index int) [][32]byte {
-	proof := [][32]byte{}
-
-	for index < len(mt)-1 {
-		siblingIndex := index ^ 1
-		proof = append(proof, mt[siblingIndex])
-		index = (len(mt) + 1 + index) / 2
-	}
-
-	return proof
-}
-
-func (mt MerkleTree) verifyProof(proof [][32]byte, leaf [32]byte) bool {
-	computedHash := leaf
-	for _, p := range proof {
-		computedHash = hashPair(computedHash, p)
-	}
-	return bytes.Compare(computedHash[:], mt[len(mt)-1][:]) == 0
-}
-
-func calculateSignedHash(rootHash [32]byte, validUntil uint64) [32]byte {
-	// Equivalent to Solidity's abi.encode(bytes32, uint64)
-	data := make([]byte, 64)
-	copy(data[:32], rootHash[:])
-	binary.BigEndian.PutUint64(data[56:], validUntil)
-
-	// Keccak256 hash of the ABI encoded parameters
-	hashedEncodedParams := crypto.Keccak256(data)
-
-	// Prepare the Ethereum signed message
-	prefix := []byte("\x19Ethereum Signed Message:\n32")
-	ethMsg := append(prefix, hashedEncodedParams...)
-
-	// Final Keccak256 hash
-	return crypto.Keccak256Hash(ethMsg)
-}
-
-func generateSignatures(t *testing.T, signers []Signer, signedHash [32]byte) [][]byte {
-	signatures := make([][]byte, len(signers))
-	for i, signer := range signers {
-		signature, err := crypto.Sign(signedHash[:], signer.privateKey)
-		require.NoError(t, err)
-
-		// Adjust the v value, we need to readd 27.
-		// ref: https://github.com/ethereum/go-ethereum/blob/b590cae89232299d54aac8aada88c66d00c5b34c/crypto/signature_nocgo.go#L90
-		v := signature[crypto.RecoveryIDOffset]
-		require.True(t, v >= 0 && v <= 3, "v should be between 0 and 3")
-		signature[crypto.RecoveryIDOffset] += 27
-
-		signatures[i] = signature
-	}
-	return signatures
-}
-
 func TestGetProof(t *testing.T) {
 	leaves := [][32]byte{
 		crypto.Keccak256Hash([]byte("leaf1")),
@@ -1031,26 +738,26 @@ func TestGetProof(t *testing.T) {
 
 	fmt.Printf("Leaves: %v\n", leaves)
 
-	tree, err := newMerkleTree(leaves)
+	tree, err := NewMerkleTree(leaves)
 	require.NoError(t, err)
 
 	fmt.Printf("Tree: %v\n", tree)
 
 	for i, leaf := range leaves {
-		proof := tree.getProof(i)
+		proof := tree.GetProof(i)
 
 		fmt.Printf("Proof for leaf %d: %v\n", i, proof)
 
 		// Verify the proof
-		isValid := tree.verifyProof(proof, leaf)
+		isValid := tree.VerifyProof(proof, leaf)
 		require.True(t, isValid, "Proof should be valid for leaf %d", i)
 
 		// Verify the proof manually
 		computedHash := leaf
 		for _, p := range proof {
-			computedHash = hashPair(computedHash, p)
+			computedHash = HashPair(computedHash, p)
 		}
-		require.Equal(t, tree.getRoot(), computedHash, "Computed root should match tree root for leaf %d", i)
+		require.Equal(t, tree.GetRoot(), computedHash, "Computed root should match tree root for leaf %d", i)
 	}
 }
